allow repo access by default on dotcom
Sourcegraph.com only has public repos, so it is safe to allow access to all repositories by all users.

This now-deleted code was originally added in https://github.com/sourcegraph/sourcegraph/pull/26345 to support private repositories on dotcom. We do not currently support private repositories on dotcom anymore, and when we do add back that feature, we can figure out how best to do it and will benefit from fewer special-cases for dotcom.