chore(svelte): Simplify fuzzy finder implementation
Currently the fuzzy finder filters and ranks results received from the
server locally. This was done to improve performance since local
filtering is much faster.
However it can introduce inconsistencies (as reported) because the local
filtering logic works differently that the one on the server. That means
the shown results is depedent on on the local cache, which is not
obvious to the user.

This commit removes the client side filtering and ranking and instead
relies only on the server for this. This makes things more conisten and
predicatable, at the expense of being a little slower. However it still
feels quite fast to me.

Note that I didn't implement some aspects due to the limitations of the
GraphQL-based search API:
- No match highlighting. Personally I didn't miss is it so far. I don't
  know if the highlighting actually provides a lot of value.
- No total result count. It seems since we are adding `count:50`, the
  server cannot actually give use an approximate total count. But we
  should probably still convey somehow that we are limiting results to
  the top 50.