support fast, simple `sg start single-program` for local dev
This makes it easier to run Sourcegraph in local dev by compiling a few key services (frontend, searcher, repo-updater, gitserver, and worker) into a single Go binary and running that.

Compared to `sg start` (which compiles and runs ~10 services), it's much faster to start up, has less log spam, and rebuilds faster. It is slower for changes to `frontend` because it needs to link in more code on each recompile.

This is only intended for local dev as a convenience. There may be different behavior in this mode that could result in problems when your code runs in the normal deployment. Usually our e2e tests should catch this, but to be safe, you should run in the usual mode if you are making sensitive cross-service changes.

Partially reverts "svcmain: Simplify service setup (#61903)" (commit 95410322925a4d78883290e01095086d11239877).