remove old and unused Cody search page
This page (at `/cody/search`) let you type in a natural-language description of what to search for, and then it used an LLM to rewrite that to Sourcegraph search query syntax. This was enabled using the `cody-web-search` feature flag, which has been disabled on dotcom for a while and was never documented or shared with customers, so it's safe to remove.

It was a cool idea, but we decided to focus on making Cody's search in the editor better, and we haven't seriously improved this for ~12 months.