chore(svelte): Properly route to revision agnostic pages
There are two kinds of repository pages: Those who depend on the
revision possibly present in the URL and those who are not.

This commit re-arranges the routes to better distinguish between these
two types. It also adds a visual indication to the commits that the
history at a specific revision is shown. Additionally it removes the
revision from links to revision agnostic pages.

Overall this should make it clearer to the user when and when not the
revision in the URL is relevant.

 Note: I originally added a redirect for revision agnostic pages that
 would remove the revision if it was present in the URL. I decided to
 not do that for now because that redirect introduces an additional
 history entry which makes it difficult to navigate back.