search: fix regexp toggle if regexp is default
I noticed that the regexp toggle doesn't work anymore if the
`"search.defaultPatternType": "regexp"`.

We also append `patternType:keyword` in that case which I don't think we
want, because we have an UI element to indicate that keyword search is
active.

The question I don't know how to answer is: What should happen if
regexp is the default and the user toggles keyword search off. Should we
go back to "regexp" or to "standard"?

Test plan:
manual testing with default pattern type set to "keyword" and "regexp".