make Cody and Code Search global navbar items one-click (#63514)
I found it annoying that it took 2 clicks to go to Cody chat: click
**Cody** in the global navbar, then click **Web Chat**. Users rarely
will visit their dashboard, and that's associated with their user
anyway, so it should be under their user. Same 2-click annoyance applies
to going to Code Search.

Moves some other links:

- Contexts, Search Jobs, Code Monitors, Notebooks --> moved to "Tools"
global navbar item
- Cody dashboard --> moved to user nav item (in top right)


![image](https://github.com/sourcegraph/sourcegraph/assets/1976/6b87ad4a-cb8a-4706-8954-f79e588803cd)

![image](https://github.com/sourcegraph/sourcegraph/assets/1976/b971c2f2-5965-4cd9-b8ac-17dc0ff9921d)

![image](https://github.com/sourcegraph/sourcegraph/assets/1976/933865d9-105f-406f-b88e-62e370896cfd)


## Test plan

With minimize nav on and off, and with all states of dotcom and
user-has-Cody-enabled, ensure matches expectations. The expectations are
codified in the tests/

## Changelog

- Code Search and Cody now are one-click links in the global navbar.
Other features are in the new "Tools" menu: search contexts, code
monitors, search jobs, and notebooks. Your Cody dashboard is linked from
your user menu (in the top right).