chore(svelte): Hide cody nav entries and upsell banner when cody isn't enabled (#63463)
Contributes to srch-529

This commit refactors the main navigation and search home page code to
make it more configurable. In particular we now only show navigation
entries for features that are enabled (as determined by
`window.context`) and only show the cody upsell banner when cody is
enabled.

I extended the dev HTML template and .env files to support this.
## Test plan

Manual testing.