fix(search): Token decoration in keyword-enabled query input (#63543)
Fixes srch-646

This commit fixes how tokens are decorated in the query input. The
"chip" style needs to have lower presedence than the syntax highlighting
so that it doesn't get split across the individual parts of e.g. a repo
filter.

This commit also removes the activeFilter extension. It was conflicting
a bit with the filter decoration. This should be fine since it wasn't
used for quite a while in the main search input either.

I also noticed an issue with the size of the CodeMirror input in the
saved searches form. I couldn't enter any query actually because the
input was too "narrow". Adding `flex:1` fixes that.


## Test plan

Manual testing.


![2024-06-28_13-12](https://github.com/sourcegraph/sourcegraph/assets/179026/4aff056c-2a18-4a41-9033-94d642c0858b)