make `<Label>` element focus & select at end to match native label behavior
When our custom `<Label>` element is labelling a contenteditable element (like a CodeMirror editor), it previously just called `element.focus()`. This focused the element but put the cursor at the start. The native and expected behavior of `<label for="...">` is to focus and put the cursor at the end.

Also fix an issue where this code path would be followed even if the label had no ID, which could result in it erroneously selecting an element with an empty `aria-labelledby` attribute.