completions: Add API version 2 with support for delta completions
Our completions API was modeled after Anthropics V1 API. Their API always returned
the whole completion string generated so far. This is a bit easier for the client,
but very wasteful, as the amount of bytes transferred is quadratic.

This PR adds a new V2 API endpoint that will return only the text that has been generated
since the last event. For that, the request has to include `?api-version=2` and for streaming
responses, the `deltaText` field will be populated.

I iterated on this a few times, and here's some other things I've tried, that I ultimately didn't settle on:

- Add a `"delta": true` field to the response, to indicate that `"completion"` contains only the delta. I didn't like this as much, as it's easy to forget to check for this, there's no type-safety in TS that can guard against it on the client side, and it sends an additional `len(,"delta":true)` bytes over the wire for every chunk.
- Make the client understand that `v2` always means "delta" implicitly, this feels easy to forget as well, and it would require a lot of extra plumbing in the client to also pass along "this response is from a request for the V2 API".

Olaf and I talked about this further, and we decided that ultimately, we should probably just implement an OpenAI compatible endpoint.
This seems to be what most in the industry are doing (see Ollama), and they most likely spend way more thoughts on the API design than
we did.
This is a larger change though, so we want to make this incremental improvement first.

Client support for V2 will still need to be implemented for this.

Test plan:

Tried the endpoint locally, for V2 requests I get delta responses. For other requests, I still get the
old behavior of returning the whole completion so far.
Attributions work again after adjusting the implementation to work with deltas.

Changelog:

- Added a completions V2 API endpoint that supports returning completion deltas to reduce bandwidth usage.