feature(cody): Data types for encoding LLM Model config via Sourcegraph site config (#63533)
This PR is part of the epic space opera that is [Server-side Cody Model
Selection](https://linear.app/sourcegraph/project/server-side-cody-model-selection-cca47c48da6d)
and friends, such as [Cody with Self-hosted
Models](https://linear.app/sourcegraph/project/cody-with-self-hosted-models-39edcf552a64).
It doesn't add any user-facing functionality, but rather introduces the
foundational data types and conversion logic so we can integrate the new
"LLM Model Configuration" code into the Sourcegraph instance.

## Overview

The `internal/modelconfig` package defines the data types and validation
logic for LLM Model Configuration. The goal is to provide an
ultra-flexible way for a Sourcegraph backend to represent LLM
configuration data, so that it can rely on it for registering its own
completions API endpoints as well as send that data to Cody clients so
that they can understand what LLM models the Sourcegraph backend
supports.

To date, all of this LLM configuration data has been bundled into the
`types.ModelConfiguration` data type. (`internal/modelconfig/types`) It
contains all of the "Providers", "Models", etc. that the Sourcegraph
instance understands.

However, this data type is not easy to manipulate. It is designed to
make it easy to look up data when we need to process an LLM completion
request, or to easily consumable by Cody clients. But there isn't a good
way for a Sourcegraph admin to "modify" or "patch" that object.

This PR introduces a new data type that is conceptually the same -- a
bunch of LLM model configuration -- but is meant to be expressed by a
Sourcegraph administrator manually configuring how Cody should work on
their instance.

### Introducing `SiteModelConfiguration`.

Inside `internal/modelconfig/types/documents.go` we define a new type,
`SiteModelConfiguration`. This is intended to be encoded directly into
the site configuration (`site.schema.json`) and be reasonably easy for a
non-expert to configure.

@slimsag and I spent a lot of time talking through the various scenarios
that we want to enable, and various tradeoffs for how to represent the
data. This is what we wound up with and is able to serve all known
requirements. (As far as we know for now. I hope. 😅)

While it isn't acceptable as a valid form of documentation (which I'll
produce once the end-to-end of this is all working, and I can show how
things come together), I'd suggest you take a look at
`internal/modelconfig/validation_test.go` to understand how to think
about this new data type. (As well as all the doc comments on the
`SiteModelConfiguration` type itself.)

### Converting the existing Site Configuration

In order to provide some more context for how we expect to _use_ these
data types, I've included some related functionality in this PR as well.

This PR adds a new `cmd/frontend/internal/modelconfig` package. This is
where we will put the Sourcegraph instance-specific logic for reading,
updating, and serving the LLM model configuration data.

For now, it just contains utility types and one meaningful function:

```go
// convertCompletionsConfig converts the supplied Completions configuration blob (the Cody Enterprise configuration data)
// into the newer types.SiteModelConfiguration structure.
//
// Assumes that the supplied completions object is valid, and contains all the required settings. e.g. the site admin
// can leave some things blank, but `conf/computed.go`'s `GetCompletionsConfig()` will fill the Endpoint and related
// fields with meaingful defaults.
func convertCompletionsConfig(completionsCfg *conftypes.CompletionsConfig) *types.SiteModelConfiguration {
```

In order to refactor our current completions APIs to use the new
`modelconfig` data types, we first need to safely read existing Cody
Enterprise configuration data and represent it in this new data
structure.

The `convertCompletionsConfig` function loads the `CompletionsConfig`
data in the site configuratg (i.e. [all this
stuff](https://sourcegraph.com/docs/cody/clients/enable-cody-enterprise)
and converts it into a `SiteModelConfiguration`.

This part is _really important_ to get right, since a mistake here would
just be a regression for any Cody Enterprise customer that upgrades
their Sourcegraph instance. (After I continue refactoring the
completions API codepath.)

We'll be able to do more meaningful testing as we land the next few PRs,
if you look at
`cmd/frontend/internal/modelconfig/siteconfig_completions_test.go` you
can get a sense for how the existing Cody Enterprise configuration data
gets surfaced in the new data types.

## Next Steps

After this PR lands, the next step is to continue building out the
`cmd/frontend/internal/modelconfig` package. Specifically, to expose the
`types.ModelConfiguration` data so that the Completions APIs
(`cmd/frontend/internal/completions`) can actually _use_ this data for
making LLM requests, either via BYOK or to Cody Gateway.

This will involve some super-annoying to write code that "merges" the
`types.ModelConfiguration` data that we'll fetch from Cody Gateway
and/or is embedded in the Sourcegraph instance with the Sourcegraph
admin supplied `types. types.SiteModelConfiguration`.

And... FINALLY!!!! at long last, once _that_ is done, it's just a matter
of updating the Completions handlers to use these new data types, and
enjoy the strongly typed and easily extendable
`ServerSideProviderConfig` configuration objects.

## Test plan

Lots of new unit tests, but none of this is surfaced to end users. YET.