Add GenericProviderConfig.ServiceName (#63626)
While trying to use the `types.ModelConfiguration` data to configure
completion providers, it became clear I missed something.

We expose the `types.GenericProviderConfig` as a bare-bones way to
express LLM provider configuration, and it does great. But just knowing
the (endpoint, access token) isn't enough to actually serve requests.

We need to know the _shape_ of API requests. e.g. if we should pass that
endpoint/access to an OpenAI-compatible, Fireworks-compatible, or
Google-compatible API provider.

> NOTE: An alternative approach would be to not add this field to
`GenericProviderConfig` entirely, and instead add `AnthropicProvider
*GenericProviderConfig`, `GoogleProvider *GenericProviderConfig`, etc.
to the `ServerSideProviderConfig`. And essentially have the "service
name" be implicit based on which field of `ServerSideProviderConfig` you
found the object.
>
> There are advantages to that approach, but this seemed a bit cleaner
IMHO. WDYT?

## Test plan

Added new unit tests

## Changelog

NA