fetch NamespaceSelector namespaces from GraphQL (`viewer { affiliatedNamespaces }`) (#63592)
Previously, the Batch Changes namespace selector looked at the user's
settings that were loaded on the initial page load. This bypassed the
GraphQL API, which should be the source of truth for "what namespaces
can a user write to" (i.e., their user account and all orgs they are a
member of), and introduced a reliance on non-type-safe `jscontext` data.
Also, adding any other features to this NamespaceSelector would require
changing the data fetched in `jscontext`, so it was brittle.

I intend to use the NamespaceSelector for more things that need these
missing features and a more solid foundation (such as the ability to
transfer saved searches to a different owner), so I started out by
refactoring this to access data from the GraphQL API. I added a new
`viewer { affiliatedNamespaces }` GraphQL API field to support this.

Also adds tests and a storybook for this NamespaceSelector component.

## Test plan

Test plan:

1. Go to the Batch Changes global create page. Create a batch change.
Ensure it works if you use the initial/default selected namespace (your
user account), and if you change it. Ensure the created batch change
exists in the intended namespace.
2. Do the same, but from a user or org batch changes create page. Ensure
the NamespaceSelector's initial option is the namespace you are
visiting, and that creation succeeds.