Add ModelRef.APIVersionID() (#63643)
The `ModelRef` encodes three pieces of information to represent the
unique identity of the model. So far, we've only needed to pull out the
`ProviderID` and `ModelID`. But now that I'm working on reading this
data for the completion APIs, it's time to add the `APIVersionID`
function too.

Note that these functions should really return an `(string, error)`
since it's possible the `ModelRef` they are operating on is invalid. But
as-is things will explode loudly enough it won't be too difficult to
detect, and leaving the function return type as `string` makes it _much_
easier to use.

I also added a comment about something I'll address when I get to adding
support for fetching new models dynamically from Cody Gateway. Right now
there is a validation check that probably should exist but doesn't, and
we cannot add it just yet.

## Test plan

Added unit tests.

## Changelog

NA