add new "modelConfiguration" schema to site config
User-facing

This PR adds a new top-level field to the site configuration `"modelConfiguration"`
which is intended to replace the old `"completions"` field. For now, this is 100%
opt-in (beta feature) and customers should only use this new field if they have
talked with us to confirm it should work in their case.

Today, this configuration is unused, but in future PRs it will actually be used (planned for the 5.5.0 release.)

Additionally, `"modelConfiguration"` acts as a feature-flag. If it is set
and not `null`, then Cody will enable this whole new model configuration system
end-to-end which involves many new components:

* Clients will make use of a new `/.api/modelconfig/supported-models.json` endpoint to query which models the server has available.
* Cody will respect this new site configuration, discovering new models from Sourcegraph's servers without an upgrade by default, etc.
* Clients will enable the "select an LLM model" dropdown menu for enterprise customers.
* The old `"completions"` configuration, if present, will be ignored if `"modelConfiguration"` is set.

Implementation notes

This schema mirrors [the new model configuration schema](https://github.com/sourcegraph/sourcegraph/tree/main/internal/modelconfig/types)
that Chris and myself have been working on tirelessly to enable a myriad
of use-cases in the near future, including enabling customers to get
support for new models without upgrading Sourcegraph, enabling multiple
models / dropdown model selector in enterprise, improved self-hosted model
support, and more.

The translation is pretty much 1:1, with a few notable aspects:

* I broke out [`GenericProviderConfig`](https://github.com/sourcegraph/sourcegraph/blob/main/internal/modelconfig/types/configuration.go#L49-L73) into distinct types for each provider.
* I represent `ClientSideProviderConfig` and `ClientSideModelConfig` _objects_ (specify multiple fields)
* I represent `ServerSideProviderConfig` and `ServerSideModelConfig` as _tagged unions_ / _discriminated unions_, i.e. where you must write a `{"type": "foo"}` field as part of the object.

Signed-off-by: Stephen Gutekanst <stephen@sourcegraph.com>